refactor(app): extract MeetingDetecting protocol
Introduce MeetingDetecting protocol with checkOnce(), isMeetingActive(),
and reset() methods. Move DetectedMeeting struct to dedicated file.
MeetingDetector now conforms to the protocol, WatchLoop depends on the
abstraction. Prepares for alternative detector implementations.
