+++ app/MeetingTranscriber/Sources/MeetingDetecting.swift
@@ -1,0 +1,44 @@
+import CoreGraphics
+import Foundation
+
+/// Represents a detected active meeting.
+struct DetectedMeeting {
+    let pattern: AppMeetingPattern
+    let windowTitle: String
+    let ownerName: String
+    let windowPID: pid_t
+    let detectedAt: Date
+
+    init(
+        pattern: AppMeetingPattern,
+        windowTitle: String,
+        ownerName: String,
+        windowPID: pid_t,
+        detectedAt: Date = Date(),
+    ) {
+        self.pattern = pattern
+        self.windowTitle = windowTitle
+        self.ownerName = ownerName
+        self.windowPID = windowPID
+        self.detectedAt = detectedAt
+    }
+}
+
+/// Protocol for meeting detection strategies.
+protocol MeetingDetecting {
+    /// Single poll: check for active meetings. Returns a meeting after confirmation threshold.
+    func checkOnce() -> DetectedMeeting?
+
+    /// Check if a previously detected meeting is still active.
+    func isMeetingActive(_ meeting: DetectedMeeting) -> Bool
+
+    /// Reset confirmation counters and start cooldown for the given app.
+    func reset(appName: String?)
+}
+
+extension MeetingDetecting {
+    // swiftlint:disable:next unused_declaration
+    func reset() {
+        reset(appName: nil)
+    }
+}
